requirements.txt
streamlit==1.32.0

langchain==0.1.16
langchain-community==0.0.34
langchain-text-splitters==0.0.1

faiss-cpu==1.12.0
pypdf==4.2.0

python-dotenv==1.0.1
tiktoken==0.6.0

sentence-transformers==2.6.1

google-genai==0.3.0